modules/location-producer/requirements.txt
aniso8601==7.0.0
attrs==19.1.0
Click==7.0
GeoAlchemy2==0.8.4
grpcio==1.41.1
grpcio-reflection==1.41.1
grpcio-tools==1.30.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.0.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
protobuf==3.19.1
psycopg2==2.9.1
psycopg2-binary==2.8.5
pycodestyle==2.8.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
Shapely==1.7.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==0.16.1
kafka-python>=2.0.2
